Addtion of an example for the median data function:
First, addition of new argument to let the user providing the data directly instead of the path (path solution is still working).

Second: creation of a toy example added to the documentation of median_data_computation

Third: addition of example data. Creation of small dataset and export into the package data folder. Need to complement the documentation.